requirements.txt
fastapi
uvicorn
motor
pydantic[email]
openai
python-dotenv
langchain
langchain-core
langchain-openai
langchain_community

pyautogen

langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langgraph-cli


tavily-python
wikipedia
trustcall
